requirements.txt
requests
rich
PyYAML
beautifulsoup4
sseclient
playwright
google-genai
esprima
tinycss
coverage
lxml
Flask
flask-sock
simple-websocket
gunicorn
pandas
